Paranthesize assert statements (#208)
* paranethesize asserts

* Update test_keras_mirrored.py
